5/11 - Updates to the interface. Edited the Calculator functions parameters to double from int. Added to the MainActivity class to incorporate the calculator functionality.

+++ app/src/main/java/com/dtchin2/calculator_app/calculator/Calculator.java
@@ -2,38 +2,38 @@ package com.dtchin2.calculator_app.calculator;
 
 public class Calculator {
 
-    private int firstNumber;
-    private int secondNumber;
+    private double firstNumber;
+    private double secondNumber;
 
-    public double add(int firstNumber, int secondNumber){
+    public double add(double firstNumber, double secondNumber){
         return firstNumber + secondNumber;
     }
 
-    public double subtract(int firstNumber, int secondNumber){
+    public double subtract(double firstNumber, double secondNumber){
         return firstNumber - secondNumber;
     }
 
-    public double multiply(int firstNumber, int secondNumber){
+    public double multiply(double firstNumber, double secondNumber){
         return firstNumber * secondNumber;
     }
 
-    public double divide(int firstNumber, int secondNumber){
+    public double divide(double firstNumber, double secondNumber){
         return firstNumber / secondNumber;
     }
 
-    public int getFirstNumber() {
+    public double getFirstNumber() {
         return firstNumber;
     }
 
-    public void setFirstNumber(int firstNumber) {
+    public void setFirstNumber(double firstNumber) {
         this.firstNumber = firstNumber;
     }
 
-    public int getSecondNumber() {
+    public double getSecondNumber() {
         return secondNumber;
     }
 
-    public void setSecondNumber(int secondNumber) {
+    public void setSecondNumber(double secondNumber) {
         this.secondNumber = secondNumber;
     }
 }
